pyproject.toml
[tool.poetry]
name = "auth-plus-monetization"
version = "0.1.0"
description = ""
authors = ["Andrew Hanasiro <andrewkhanasiro@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pydantic = "^2.10.0"
python-dotenv = "^1.0.1"
kafka-python = "^2.0.2"
gevent = "^24.11.1"
typing-extensions = "^4.12.2"
schedule = "^1.2.2"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
pytest = "^8.3.3"
coverage = "^7.6.7"
mypy = "^1.13.0"
isort = "^5.13.2"
flake8-pyproject = "^1.2.3"
types-requests = "^2.32.0.20241016"
httpx = "^0.27.2"
responses = "^0.25.3"

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88

[build-system]
requires = ["poetry-core>=1.8"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = ["kafka", "pook"]
ignore_missing_imports = true
